requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
groq==0.9.0
httpx==0.27.0
psycopg2-binary==2.9.9
